requirements.txt
streamlit==1.28.2
opencv-contrib-python-headless==4.10.0.84
numpy==1.24.3
Pillow==9.5.0
matplotlib==3.7.1
torch==2.1.0
